+++ neo4j/db/summary.go
@@ -90,6 +90,8 @@ type ProfiledPlan struct {
 	Time              int64
 }
 
+// StreamSummary is part of the GQL compliant notifications preview feature
+// (see README on what it means in terms of support and compatibility guarantees)
 type StreamSummary struct {
 	HadRecord bool
 	HadKey    bool
@@ -104,18 +106,51 @@ type Notification struct {
 	Category    string
 }
 
+// GqlStatusObject represents a GqlStatusObject generated when executing a statement.
+// A GqlStatusObject can be visualized in a client pinpointing problems or other information about the statement.
+// Contrary to failures or errors, GqlStatusObjects do not affect the execution of the statement.
+//
+// GqlStatusObject is part of the GQL compliant notifications preview feature
+// (see README on what it means in terms of support and compatibility guarantees)
 type GqlStatusObject struct {
 	// Deprecated: for backward compatibility with Notification.Code only.
 	Code string
 	// Deprecated: for backward compatibility with Notification.Title only.
-	Title             string
-	GqlStatus         string
+	Title string
+	// GqlStatus returns the GQLSTATUS.
+	// The following GQLSTATUS codes denote codes that the driver will use for
+	// polyfilling (when connected to an old, non-GQL-aware server). Further, they may be used by servers
+	// during the transition-phase to GQLSTATUS-awareness.
+	//   - "01N42" (warning - unknown warning)
+	//   - "02N42" (no data - unknown subcondition)
+	//   - "03N42" (informational - unknown notification)
+	//   - "05N42" (general processing exception - unknown error)
+	//
+	// This means these codes are not guaranteed to be stable and may change in future versions.
+	GqlStatus string
+	// StatusDescription returns a description of the status.
 	StatusDescription string
-	Position          *InputPosition
-	Classification    string
-	Severity          string
-	DiagnosticRecord  map[string]any
-	IsNotification    bool
+	// Position returns the position in the statement where this status points to.
+	// Not all statuses have a unique position to point to and in that case the position would be set to nil.
+	//
+	// Only Notifications (see IsNotification) have a meaningful position.
+	Position *InputPosition
+	// Classification returns the mapped Classification of this status.
+	// If the Classification is not a known value, Classification returns notifications.UnknownClassification.
+	//
+	// Only Notifications (see IsNotification) have a meaningful classification.
+	Classification string
+	// Severity returns the mapped Severity of this status.
+	// If the Severity is not a known value, Severity returns notifications.UnknownSeverity.
+	//
+	// Only Notifications (see IsNotification) have a meaningful severity.
+	Severity string
+	// DiagnosticRecord returns further information about the status for diagnostic purposes.
+	DiagnosticRecord map[string]any
+	// IsNotification returns true if this status is a Notification.
+	//
+	// Only some statuses are Notifications.
+	IsNotification bool
 }
 
 // InputPosition contains information about a specific position in a statement
