+++ neo4j/internal/bolt/bolt4_test.go
@@ -420,7 +420,6 @@ func TestBolt4(outer *testing.T) {
 		type testCase struct {
 			description string
 			MinSev      notifications.NotificationMinimumSeverityLevel
-			//lint:ignore SA1019 NotificationDisabledCategories is supported at least until 6.0
 			DisCats     notifications.NotificationDisabledCategories
 			ExpectError bool
 			Method      string
@@ -440,7 +439,6 @@ func TestBolt4(outer *testing.T) {
 				},
 				testCase{
 					description: "disabled categories",
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
 					DisCats:     notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectError: true,
 					Method:      s,
@@ -448,14 +446,12 @@ func TestBolt4(outer *testing.T) {
 				testCase{
 					description: "warning minimum severity and disabled categories",
 					MinSev:      notifications.WarningLevel,
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
 					DisCats:     notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectError: true,
 					Method:      s,
 				},
 				testCase{
 					description: "disable no categories",
-					//lint:ignore SA1019 DisableNoCategories is supported at least until 6.0
 					DisCats:     notifications.DisableNoCategories(),
 					ExpectError: true,
 					Method:      s,
