+++ neo4j/internal/bolt/bolt5_test.go
@@ -563,9 +563,8 @@ func TestBolt5(outer *testing.T) {
 	outer.Run("with notifications", func(inner *testing.T) {
 		warningSev := "WARNING"
 		type testCase struct {
-			description string
-			MinSev      notifications.NotificationMinimumSeverityLevel
-			//lint:ignore SA1019 NotificationDisabledCategories is supported at least until 6.0
+			description     string
+			MinSev          notifications.NotificationMinimumSeverityLevel
 			DisCats         notifications.NotificationDisabledCategories
 			ExpectedMinSev  *string
 			ExpectDisCats   bool
@@ -586,17 +585,15 @@ func TestBolt5(outer *testing.T) {
 					Method:         s,
 				},
 				testCase{
-					description: "disabled categories",
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
+					description:     "disabled categories",
 					DisCats:         notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectDisCats:   true,
 					ExpectedDisCats: []any{"UNSUPPORTED", "GENERIC"},
 					Method:          s,
 				},
 				testCase{
-					description: "warning minimum severity and disabled categories",
-					MinSev:      notifications.WarningLevel,
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
+					description:     "warning minimum severity and disabled categories",
+					MinSev:          notifications.WarningLevel,
 					DisCats:         notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectDisCats:   true,
 					ExpectedDisCats: []any{"UNSUPPORTED", "GENERIC"},
@@ -604,8 +601,7 @@ func TestBolt5(outer *testing.T) {
 					Method:          s,
 				},
 				testCase{
-					description: "disable no categories",
-					//lint:ignore SA1019 DisableNoCategories is supported at least until 6.0
+					description:     "disable no categories",
 					DisCats:         notifications.DisableNoCategories(),
 					ExpectDisCats:   true,
 					ExpectedDisCats: []any{},
@@ -672,7 +668,6 @@ func TestBolt5(outer *testing.T) {
 		type testCase struct {
 			description string
 			MinSev      notifications.NotificationMinimumSeverityLevel
-			//lint:ignore SA1019 NotificationDisabledCategories is supported at least until 6.0
 			DisCats     notifications.NotificationDisabledCategories
 			ExpectError bool
 			Method      string
@@ -692,7 +687,6 @@ func TestBolt5(outer *testing.T) {
 				},
 				testCase{
 					description: "disabled categories",
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
 					DisCats:     notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectError: true,
 					Method:      s,
@@ -700,14 +694,12 @@ func TestBolt5(outer *testing.T) {
 				testCase{
 					description: "warning minimum severity and disabled categories",
 					MinSev:      notifications.WarningLevel,
-					//lint:ignore SA1019 DisableCategories is supported at least until 6.0
 					DisCats:     notifications.DisableCategories(notifications.Unsupported, notifications.Generic),
 					ExpectError: true,
 					Method:      s,
 				},
 				testCase{
 					description: "disable no categories",
-					//lint:ignore SA1019 DisableNoCategories is supported at least until 6.0
 					DisCats:     notifications.DisableNoCategories(),
 					ExpectError: true,
 					Method:      s,
@@ -1937,9 +1929,8 @@ func TestBolt5(outer *testing.T) {
 	for _, test := range txTimeoutTestCases {
 		outer.Run(test.description, func(t *testing.T) {
 			tx := internalTx5{timeout: test.input}
-
-			minor := 0
-			actual, ok := tx.toMeta(logger, "", minor)["tx_timeout"]
+			version := db.ProtocolVersion{Major: 5, Minor: 0}
+			actual, ok := tx.toMeta(logger, "", version)["tx_timeout"]
 			if test.omitted {
 				if ok {
 					t.Errorf("tx_timeout was present but should be omitted")
