+++ neo4j/internal/db/connection.go
@@ -59,28 +59,24 @@ type TxConfig struct {
 }
 
 type NotificationConfig struct {
-	MinSev notifications.NotificationMinimumSeverityLevel
-	//lint:ignore SA1019 NotificationDisabledCategories is supported at least until 6.0
+	MinSev  notifications.NotificationMinimumSeverityLevel
 	DisCats notifications.NotificationDisabledCategories
 	DisClas notifications.NotificationDisabledClassifications
 }
 
-func (n *NotificationConfig) ToMeta(meta map[string]any, minor int) {
+func (n *NotificationConfig) ToMeta(meta map[string]any, version db.ProtocolVersion) {
 	if n.MinSev != notifications.DefaultLevel {
 		meta["notifications_minimum_severity"] = string(n.MinSev)
 	}
 
 	disabledKey := "notifications_disabled_categories"
-	if minor >= 5 {
+	if version.Minor >= 5 {
 		disabledKey = "notifications_disabled_classifications"
 	}
-	//lint:ignore SA1019 DisablesNone is supported at least until 6.0
 	if n.DisCats.DisablesNone() || n.DisClas.DisablesNone() {
 		meta[disabledKey] = make([]string, 0)
 	} else {
-		//lint:ignore SA1019 DisabledCategories is supported at least until 6.0
 		if len(n.DisCats.DisabledCategories()) > 0 {
-			//lint:ignore SA1019 DisabledCategories is supported at least until 6.0
 			meta[disabledKey] = n.DisCats.DisabledCategories()
 		}
 		if len(n.DisClas.DisabledClassifications()) > 0 {
