+++ neo4j/notifications_examples_test.go
@@ -28,7 +28,6 @@ func ExampleConfig_disableNoCategories() {
 	ctx := context.Background()
 	driver, err := NewDriverWithContext(getUrl(), getAuth(), func(config *Config) {
 		// makes the server return all notification categories
-		//lint:ignore SA1019 NotificationsDisabledCategories is supported at least until 6.0
 		config.NotificationsDisabledCategories = notifications.DisableNoCategories()
 	})
 	handleError(err)
@@ -53,7 +52,6 @@ func ExampleSessionConfig_disableNoCategories() {
 	session := driver.NewSession(ctx, SessionConfig{
 		// makes the server return all notification categories
 		// this overrides the driver level configuration of the same name
-		//lint:ignore SA1019 DisableNoCategories is supported at least until 6.0
 		NotificationsDisabledCategories: notifications.DisableNoCategories(),
 	})
 
@@ -74,7 +72,6 @@ func ExampleConfig_disableSomeCategories() {
 	ctx := context.Background()
 	driver, err := NewDriverWithContext(getUrl(), getAuth(), func(config *Config) {
 		// makes the server return all notification categories but deprecations
-		//lint:ignore SA1019 NotificationsDisabledCategories and DisableCategories are supported at least until 6.0
 		config.NotificationsDisabledCategories = notifications.DisableCategories(notifications.Deprecation)
 	})
 	handleError(err)
@@ -100,7 +97,6 @@ func ExampleSessionConfig_disableSomeCategories() {
 	session := driver.NewSession(ctx, SessionConfig{
 		// makes the server return all notification categories but deprecations
 		// this overrides the driver level configuration of the same name
-		//lint:ignore SA1019 NotificationsDisabledCategories and DisableCategories are supported at least until 6.0
 		NotificationsDisabledCategories: notifications.DisableCategories(notifications.Deprecation),
 	})
